+++ packages/ui/src/components/Radio/__stories__/Template.stories.tsx
@@ -4,7 +4,6 @@ import { Radio } from '..'
 export const Template: StoryFn<typeof Radio> = args => <Radio {...args} />
 
 Template.args = {
-  checked: false,
   label: 'Label 1',
   disabled: false,
   name: 'basic',