+++ packages/ui/src/components/Table/index.tsx
@@ -1,6 +1,5 @@
 import { useTheme } from '@emotion/react'
 import styled from '@emotion/styled'
-import type { ReactNode } from 'react'
 import { forwardRef } from 'react'
 import { Body } from './Body'
 import { Cell } from './Cell'
@@ -11,6 +10,7 @@ import { Row } from './Row'
 import { SelectBar } from './SelectBar'
 import { SkeletonRows } from './SkeletonRows'
 import { TableProvider, useTableContext } from './TableContext'
+import type { TableProviderProps } from './TableContext'
 import { EXPANDABLE_COLUMN_SIZE, SELECTABLE_CHECKBOX_SIZE } from './constants'
 import type { ColumnProps } from './types'
 
@@ -59,11 +59,24 @@ const StyledTable = styled('table', {
   }
   `}
 `
+// type OptionalKeys<T> = {
+//   [K in keyof T]: {} extends Pick<T, K> ? K : never
+// }[keyof T]
 
-type TableProps = {
+// type OptionalOnly<T> = Pick<T, OptionalKeys<T>>
+
+// TODO: Get type optional from omit values
+type TableProps = Omit<
+  TableProviderProps,
+  | 'selectable'
+  | 'loading'
+  | 'bordered'
+  | 'stripped'
+  | 'autoCollapse'
+  | 'columns'
+  | 'expandButton'
+> & {
   selectable?: boolean
-  columns: ColumnProps[]
-  children: ReactNode
   /**
    * Set it to true if you want to display a placeholder during loading
    * */
@@ -75,6 +88,8 @@ type TableProps = {
    * Auto collapse is collapsing expandable row when another is expanding
    * */
   autoCollapse?: boolean
+  expandButton?: boolean
+  columns: ColumnProps[]
 }
 
 export const BaseTable = forwardRef<HTMLTableElement, TableProps>(