+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 1.82.1
+
+### Patch Changes
+
+- [#4615](https://github.com/scaleway/ultraviolet/pull/4615) [`39e17a6`](https://github.com/scaleway/ultraviolet/commit/39e17a6842f8bdf58f2500b5867937c96e0c4b68) Thanks [@matthprost](https://github.com/matthprost)! - Fix `<List />` when disabled state the row doens't change color on hover
+
+- [#4603](https://github.com/scaleway/ultraviolet/pull/4603) [`e94eaad`](https://github.com/scaleway/ultraviolet/commit/e94eaad0be3138679cee9743b13335c32f71d3a2) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `@types/react` to `19.0.1`.
+  Updated dependency `@types/react-dom` to `19.0.2`.
+
+- [#4609](https://github.com/scaleway/ultraviolet/pull/4609) [`e14c76d`](https://github.com/scaleway/ultraviolet/commit/e14c76d530ba5be85602c07e8351ecddbc07ac39) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `@types/react` to `19.0.2`.
+
+- [#4619](https://github.com/scaleway/ultraviolet/pull/4619) [`79a5b3a`](https://github.com/scaleway/ultraviolet/commit/79a5b3ad7a302a43c816081001f2f96c7556c33c) Thanks [@philibea](https://github.com/philibea)! - fix checkbox event on List & Table component
+
+- Updated dependencies [[`e94eaad`](https://github.com/scaleway/ultraviolet/commit/e94eaad0be3138679cee9743b13335c32f71d3a2), [`e14c76d`](https://github.com/scaleway/ultraviolet/commit/e14c76d530ba5be85602c07e8351ecddbc07ac39)]:
+  - @ultraviolet/icons@3.7.1
+
 ## 1.82.0
 
 ### Minor Changes